requirements.txt
streamlit==1.22.0
tensorflow==2.12.0
numpy==1.23.5
pillow==9.5.0
opencv-python==4.5.5.64
matplotlib==3.7.1
requests==2.31.0
